Problems resolved
Fixed the indentation, added spaces around operators and fixed the file name.
